Revert "Bug 2015410 - Part 2: Fix clang-tidy warnings in js/src/builtin/temporal. r=spidermonkey-reviewers,arai" for causing Calendar.cpp related failures.
This reverts commit 1f9af96e8664f756c4238384a3714a7295e2a55b.

This reverts commit 9d0bbd803fcdebd0650f3657a3d328940a6ca47a.

+++ js/src/builtin/intl/DurationFormat.cpp
@@ -684,8 +684,7 @@ static bool InitializeDurationFormat(
                          &fractionalDigits)) {
       return false;
     }
-    dfOptions->fractionalDigits =
-        static_cast<int8_t>(fractionalDigits.valueOr(-1));
+    dfOptions->fractionalDigits = fractionalDigits.valueOr(-1);
   }
   durationFormat->setOptions(dfOptions.release());
   AddCellMemory(durationFormat, sizeof(DurationFormatOptions),
